+++ src/main.js
@@ -45,9 +45,14 @@ import Button from 'primevue/button'
 import Card from 'primevue/card'
 import Carousel from 'primevue/carousel'
 import Chip from 'primevue/chip'
+import Column from 'primevue/column'
+import ConfirmDialog from 'primevue/confirmdialog'
+import ConfirmationService from 'primevue/confirmationservice'
+import DataTable from 'primevue/datatable'
 import Dialog from 'primevue/dialog'
 import Divider from 'primevue/divider'
 import Fieldset from 'primevue/fieldset'
+import FileUpload from 'primevue/fileupload'
 import IconField from 'primevue/iconfield'
 import Image from 'primevue/image'
 import InlineMessage from 'primevue/inlinemessage'
@@ -63,8 +68,8 @@ import SplitButton from 'primevue/splitbutton'
 import TabPanel from 'primevue/tabpanel'
 import TabView from 'primevue/tabview'
 import Toast from 'primevue/toast'
-import ToggleButton from 'primevue/togglebutton'
 import ToastService from 'primevue/toastservice'
+import ToggleButton from 'primevue/togglebutton'
 import Tooltip from 'primevue/tooltip'
 
 app.use(PrimeVue, { pt: Wind })
@@ -74,9 +79,13 @@ app.component('Button', Button)
 app.component('Card', Card)
 app.component('CarouselP', Carousel)
 app.component('Chip', Chip)
+app.component('Column', Column)
+app.component('ConfirmDialog', ConfirmDialog)
+app.component('DataTable', DataTable)
 app.component('Dialog', Dialog)
 app.component('Divider', Divider)
 app.component('Fieldset', Fieldset)
+app.component('FileUpload', FileUpload)
 app.component('IconField', IconField)
 app.component('Image', Image)
 app.component('InlineMessage', InlineMessage)
@@ -93,6 +102,7 @@ app.component('TabPanel', TabPanel)
 app.component('TabView', TabView)
 app.component('Toast', Toast)
 app.component('ToggleButton', ToggleButton)
+app.use(ConfirmationService);
 app.use(ToastService)
 app.use(router)
 app.directive('tooltip', Tooltip)